+++ src/hooks/api/applications/useApplications.ts
@@ -1,4 +1,4 @@
-import useAuthSWR from "src/hooks/api/common/useAuthSWR";
+import useAuthSWR from "src/hooks/api/useAuthSWR";
 import { CommonListResponse } from "src/types/api/common";
 import { Application } from "src/types/api/application";
 import { API_ROUTES } from "src/constants/routes";
