+++ src/hooks/api/index/usePostSummary.ts
@@ -1,4 +1,4 @@
-import useAuthSWR from "src/hooks/api/common/useAuthSWR";
+import useAuthSWR from "src/hooks/api/useAuthSWR";
 import { API_ROUTES } from "src/constants/routes";
 import { PostPreview } from "src/types/api/post";
